chore(ts-update): move methods in the transformer
